ARC-778 - updated tests to reflect expected url structure
