+++ test/fixtures/api/graphql/branch-ref-nodes.json
@@ -13,7 +13,7 @@
                   }
                 ]
               },
-              "name": "TES-321-branch-name",
+              "name": "test-branch-name",
               "target": {
                 "author": {
                   "avatarUrl": "https://camo.githubusercontent.com/test-avatar",
