+++ test/safe/branch.test.ts
@@ -83,7 +83,7 @@ describe("Branch Webhook", () => {
 						id: "test-repo-id",
 						branches: [
 							{
-								createPullRequestUrl: "test-repo-url/compare/TES-123-test-ref?title=TES-123-TES-123-test-ref",
+								createPullRequestUrl: "test-repo-url/compare/TES-123-test-ref?title=TES-123%20-%20TES-123-test-ref",
 								lastCommit: {
 									author: {
 										name: "test-branch-author-name",
@@ -195,7 +195,7 @@ describe("Branch Webhook", () => {
 						id: "test-repo-id",
 						branches: [
 							{
-								createPullRequestUrl: "test-repo-url/compare/TES-123-test-ref?title=TES-123-TES-123-test-ref",
+								createPullRequestUrl: "test-repo-url/compare/TES-123-test-ref?title=TES-123%20-%20TES-123-test-ref",
 								lastCommit: {
 									author: {
 										name: "test-branch-author-name",
